+++ src/gwproxy/gwproxy.h
@@ -128,8 +128,8 @@ struct gwp_conn_pair {
 
 struct gwp_conn_slot {
 	struct gwp_conn_pair	**pairs;
-	size_t			nr;
-	size_t			cap;
+	uint32_t		nr;
+	uint32_t		cap;
 };
 
 #ifdef CONFIG_IO_URING
